% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cp_cols.R
\name{add_cp_cols}
\alias{add_cp_cols}
\title{Add columns relevant to BLE LTER Core Program}
\usage{
add_cp_cols(
  df,
  station_code_col = "station",
 
    station_source = "https://utexas.box.com/shared/static/sjt5phkdpyx9vsvpvcss461562vb5wsw.csv"
)
}
\arguments{
\item{df}{(data.frame) Data frame in question with CP data}

\item{station_code_col}{(character) Name of column containing BLE LTER station codes in supplied df. Defaults to "station".}

\item{station_source}{(character) URL or file path to Core Program stations coordinates file. I use Box's direct download feature to generate URLs for this. The source file lives at https://utexas.app.box.com/file/1092468994724?s=sjt5phkdpyx9vsvpvcss461562vb5wsw.}
}
\value{
Data with new station name, lat/lon, habitat type columns appended
}
\description{
Add columns: station name, lat/lon, habitat type, given BLE LTER Core Program station code. If there are unmatched station codes, a warning will be thrown but the rows will be preserved.
}
