% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_units.R
\name{append_units}
\alias{append_units}
\title{Append units to column names}
\usage{
append_units(meta_list, skip = NULL)
}
\arguments{
\item{meta_list}{(list) A metadata list structure returned by MetaEgress::get_meta()}

\item{skip}{(numeric) Vector of entity numbers to exclude from appending units. E.g. raw data entities might need to be archived as-is.}
}
\description{
Append abbreviated units to attribute names
}
