% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classification.R
\name{check_classification}
\alias{check_classification}
\title{Checks classification mismatches between a dataframe and the WoRMS database
This function verifies the taxonomic classification of rows in a dataframe against data from the WoRMS API based on AphiaID. It iterates over each unique AphiaID in the dataframe, fetches the corresponding classification from WoRMS, and compares it against the dataframe values. 
Mismatches are collected and printed. If no dataframe is provided or the input is not a dataframe, the function will stop with an error.}
\usage{
check_classification(input_dataframe, file_path)
}
\arguments{
\item{input_dataframe}{A dataframe with at least one column named `aphia_ID` and other columns corresponding to taxonomic levels (kingdom, phylum, class, order, family, genus, species). Taxa levels columns must be lower case.}

\item{file_path}{A character string representing the file path where the output will be saved, e.g., "C:/Users/YourName/Documents/output.txt".}
}
\value{
Returns a list of mismatches, where each mismatch is a list containing the AphiaID and the discrepancies between the dataframe and the WoRMS data. Each mismatch specifies the dataframe value and the WoRMS value for each differing taxonomic level.
}
\description{
Checks classification mismatches between a dataframe and the WoRMS database
This function verifies the taxonomic classification of rows in a dataframe against data from the WoRMS API based on AphiaID. It iterates over each unique AphiaID in the dataframe, fetches the corresponding classification from WoRMS, and compares it against the dataframe values. 
Mismatches are collected and printed. If no dataframe is provided or the input is not a dataframe, the function will stop with an error.
}
