% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_adc_to_edi.R
\name{compare_adc_to_edi}
\alias{compare_adc_to_edi}
\title{Compare ADC to EDI}
\usage{
compare_adc_to_edi(adc_auth_token)
}
\arguments{
\item{adc_auth_token}{(character) The authentication token for the Arctic Data Center. See https://arcticdata.io/catalog/api}
}
\description{
Reports datasets at EDI that the Arctic Data Center needs to harvest
}
