% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_stations.R
\name{correct_stations}
\alias{correct_stations}
\title{Correct station codes on certain data rows based on dates sampled.}
\usage{
correct_stations(
  df,
  station_code_col = "station",
 
    corrections_source = "https://utexas.box.com/shared/static/1uy9n875jtc7djxlckm0ogv6v3mwes6f.csv"
)
}
\arguments{
\item{df}{(data.frame) Data.frame of data with stations to be corrected}

\item{station_code_col}{(character) Name of column containing station codes}

\item{corrections_source}{(character) URL to corrections CSV file. I use a shared Box file for this. The source file is at https://utexas.app.box.com/file/1135058499563?s=1uy9n875jtc7djxlckm0ogv6v3mwes6f.}
}
\value{

}
\description{
The need for this function arose when in 2023 it was determined that in the past some stations were not labelled correctly during break-up season (june/july). Due to ice conditions, on certain dates sampling was done not at the station (within 100m of waypoint coordinates). This function takes a csv of corrections of dates, old and new station codes, and convert the old codes to new only on these dates. Note that this removes the station information columns (coordinates, etc) and one can add it back with add_cp_cols.
}
