% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp_data}
\alias{cp_data}
\title{example of BLE LTER Core Program data with various formatting issues. use to test functions in this package}
\format{
An object of class \code{data.frame} with 104 rows and 7 columns.
}
\usage{
cp_data
}
\description{
example of BLE LTER Core Program data with various formatting issues. use to test functions in this package
}
\keyword{datasets}
