% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_personnel.R
\name{export_personnel}
\alias{export_personnel}
\title{Connect to metabase and query for personnel list.}
\usage{
export_personnel(
  dbname = "ble_metabase",
  schema = "pkg_mgmt",
  dataset_ids,
  host = "10.157.18.83",
  port = 5432,
  user = NULL,
  password = NULL,
  write_to_file = TRUE,
  file_dir = getwd(),
  file_name = NULL
)
}
\arguments{
\item{dbname}{(character) name of database. Defaults to "ble_metabase".}

\item{schema}{(character) name of schema containing views. Defaults to 'pkg_mgmt'.}

\item{dataset_ids}{(numeric) Number or numeric vector of dataset IDs to query.}

\item{host}{(character) host name or IP address. Defaults to 'localhost'.}

\item{port}{(numeric) port number. Defaults to 5432.}

\item{user}{(character) (optional) username to use in connecting to database. Use to save time. If not supplied, the R console will prompt you to enter a username.}

\item{password}{(character) (optional) password to user. Use to save time. If not supplied, the R console will prompt you to enter a password.}

\item{write_to_file}{(logical) Whether to write to file. Defaults to TRUE.}

\item{file_dir}{(character) Full path of directory to write file to. Defaults to current working directory.}

\item{file_name}{(character) File name, will default to "BLE_LTER_(dataset_ids)_personnel.csv" if not specified.}
}
\value{
(data.frame) A data frame of personnel associated with specified dataset IDs as queried from metabase.
}
\description{
Connect to instance of LTER-core-metabase and query for personnel list with years associated (BLE LTER Core Program data package requirement).
}
