% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_season.R
\name{infer_season}
\alias{infer_season}
\title{Infer BLE LTER field campaign season from date or date-times}
\usage{
infer_season(data, date_col = "date_time")
}
\arguments{
\item{data}{data.frame A date.frame with date}

\item{date_col}{character Name of column with dates or date-times. Please convert this column into a Date format first (not just strings containing dates).}

\item{format}{for}
}
\value{
A vector of strings: field campaign seasons in same order as dates or date-times supplied.
}
\description{
Infer BLE LTER field campaign season from date or date-times: dates in May or earlier is under ice, June to 15th of July is break up, and 16th of July or later is open water. Do not use with mooring data.
}
