% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_datapkg.R
\name{init_datapkg}
\alias{init_datapkg}
\title{Create folder structure for a new BLE LTER data package ID
Create these following items: folder structure, R project, reusable R script from template}
\usage{
init_datapkg(dataset_id, dataset_nickname, base_path = "K:/Data")
}
\arguments{
\item{dataset_id}{(numeric) Dataset ID}

\item{dataset_nickname}{(character)}

\item{base_path}{(character) Full file path to parent directory to initiate under. Defaults to "K:/Data", which is "\\austin.utexas.edu/disk/engr/research/crwr/projects/BLE-LTER/Data" mapped to the K drive on An's PC.}
}
\value{
Nothing in R, but new folder structure and R script from template
}
\description{
Create folder structure for a new BLE LTER data package ID
Create these following items: folder structure, R project, reusable R script from template
}
