% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_additional_metadata.R
\name{insert_additional_metadata}
\alias{insert_additional_metadata}
\title{Insert additional metadata into an EML document}
\usage{
insert_additional_metadata(file, pattern = NULL, addition = NULL, after)
}
\arguments{
\item{file}{(character) Path to EML file. Function will also write to this same file.}

\item{pattern}{(character) Grep pattern. \code{grep()} will be called on the contents of the file, and the first match will be used. Defaults to "<\dataset>".}

\item{addition}{(character) Plain text of EML snippet to insert. Defaults to an "additionalMetadata" node pertaining to replication I won't paste here.}

\item{after}{(numeric) A line number to append after. Only comes into play if pattern is not found. No defaults.}
}
\value{
Nothing in the R environment, but a file with the snippet inserted where specified. File will remain the same
}
\description{
Sometime, for various reasons, additional metadata just needs to be inserted into EML documents manually. This is for quickly doing that in R. The function adds a snippet of plain text EML after a specified line number in the EML file. A pattern can also be defined for grep-ing the line number to append to (e.g. additional metadata needs to be inserted after a specific line). The defaults reflect BLE's need to add a snippet to aid replication of metadata to the Arctic Data Center that for a certain reason cannot be added via our normal means (the EML R package).
}
