% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_attributes.R
\name{rename_attributes}
\alias{rename_attributes}
\title{Rename data columns according to metadata.}
\usage{
rename_attributes(meta_list, dataset_id, entity, file_dir = getwd(),
  filename = "", x)
}
\arguments{
\item{meta_list}{(character) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{entity}{(numeric) An entity number.}

\item{file_dir}{(character) Path to directory containing flat files (data files). Defaults to current R working directory.}

\item{filename}{(character) Filename. Defaults to "", in which case the entity metadata will be read to find filename.}

\item{x}{R object. If specified, will skip looking for entity file altogether.}
}
\value{
The data table entity with renamed columns.
}
\description{
Rename data columns according to metadata.
}
