% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_cp_data.R
\name{sort_cp_rows}
\alias{sort_cp_rows}
\title{Sort Core Program data}
\usage{
sort_cp_rows(data, type)
}
\arguments{
\item{data}{(data.frame) A data.frame containing Core Program data. Columns already need to be ordered in the particular CP order. Use bleutils::order_cp_cols for this. Columns names also need to be named according to standard CP column names. Use bleutils::reaname_attributes for this.}

\item{type}{(character) Type of CP data. Choose from "water", "sediment", or "mooring". This affects the columns on which to sort data.}
}
\value{
(data.frame) The same data but ordered.
}
\description{
Sort Core Program data in preferred sort order: node, lagoon, station, date/date_time, then water_column_position if applicable
}
