% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_info.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{BLE LTER Core Program, information on core sampling stations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 8 columns.
}
\usage{
station_info
}
\description{
BLE LTER Core Program, information on core sampling stations
}
\keyword{datasets}
