% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{BLE LTER Core Program, information on core sampling stations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 8 columns.
}
\usage{
stations
}
\description{
BLE LTER Core Program, information on core sampling stations
}
\keyword{datasets}
