% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cp_cols.R
\name{update_cp_stations}
\alias{update_cp_stations}
\title{Compare the station information to "canonical" source in Box and update if needed. Meant to be run occassionally from An's PC and within the bleutils Rproject to update the version of station info that this package uses to append new columns, for example.}
\usage{

  update_cp_stations(source_file = "C:/Users/atn893/Documents/Box Sync/Beaufort LTER/Core Program/BLE_LTER_CP_Stations.xlsx",
  sheet = "lookup")
}
\arguments{
\item{source_file}{(character) Full path to canonical file used to store station codes, names, lat/lons, habitat type. Defaults to "C:/Users/atn893/Documents/Box Sync/Beaufort LTER/Core Program/BLE_LTER_CP_Stations.xlsx" which works on An's PC.}

\item{sheet}{(character) Name of Excel sheet containing lookup info. Defaults to "lookup".
importFrom readxl read_excel}
}
\description{
Compare the station information to "canonical" source in Box and update if needed. Meant to be run occassionally from An's PC and within the bleutils Rproject to update the version of station info that this package uses to append new columns, for example.
}
