% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cp_cols.R
\name{add_cp_cols}
\alias{add_cp_cols}
\title{Add columns relevant to BLE LTER Core Program
Add columns: station name, lat/lon, habitat type, given BLE LTER Core Program station code}
\usage{
add_cp_cols(df, station_code_col)
}
\arguments{
\item{df}{(data.frame) Data frame in question with CP data}

\item{station_code_col}{(character) Name of column containing BLE LTER station codes}
}
\value{
Data with new station name, lat/lon, habitat type columns appended
}
\description{
Add columns relevant to BLE LTER Core Program
Add columns: station name, lat/lon, habitat type, given BLE LTER Core Program station code
}
